Add Cold Startup prefab to play directly from gameplay scenes
Add a prefab EditorInitializer that can be place in any gameplay scene to load the persistenceManger Scene additively first.

Separate logic for initialization in a different Scene, that is actually the only one present in the Build Settings. This scene will be responsible to load PersistenceManager Scene, ask it to load Main Menu and finally unload itself.